Initial commit of handling of onPause()
onPause() is called on ICS when the back button is pressed before the task lists
is shown. With this commit, opentyrian pauses an ongoing level and waits for a key
before continuing it.

It does however not silence the music, which may be desirable as well
(to be added with the next commit)